Updated support for angular rc.6 Had to add a module for angular2-grid, replaced calls from depricated DynamicComponentLoader to componentFactoryResolver. To use in rc6 and beyond, in app NgModule import NgGridModule and the rest should load. Demo or the index under src are not updated with the new calls.
