+++ src/main.ts
@@ -1,4 +1,5 @@
-export * from "./directives/NgGrid";
-export * from "./directives/NgGridItem";
-export * from "./components/NgGridPlaceholder";
-export * from "./interfaces/INgGrid";
+export * from './directives/NgGrid';
+export * from './directives/NgGridItem';
+export * from './components/NgGridPlaceholder';
+export * from './interfaces/INgGrid';
+export * from './modules/NgGrid.module';
